add torch_networks placeholder
