+++ nn_meter/builder/nn_generator/__init__.py
@@ -1,2 +1,3 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
+from .interface import BaseOperator, BaseBlock
