+++ nn_meter/builder/nn_generator/torch_networks/operators.py
@@ -1,0 +1,7 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT license.
+import logging
+import torch
+import torch.nn as nn
+from ..interface import BaseOperator
+logging = logging.getLogger("nn-Meter")
